Move content of storage/tsdb to storage/backend(cortexproject#2262
To make code support of backends(azure, s3 etc). The contents of
storage/tsdb was moved to storage/backend
