+++ pkg/storage/backend/azure/bucket_client.go
