+++ pkg/storage/tsdb/bucket_client.go
@@ -6,10 +6,10 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/thanos-io/thanos/pkg/objstore"
 
-	"github.com/cortexproject/cortex/pkg/storage/tsdb/backend/azure"
-	"github.com/cortexproject/cortex/pkg/storage/tsdb/backend/filesystem"
-	"github.com/cortexproject/cortex/pkg/storage/tsdb/backend/gcs"
-	"github.com/cortexproject/cortex/pkg/storage/tsdb/backend/s3"
+	"github.com/cortexproject/cortex/pkg/storage/backend/azure"
+	"github.com/cortexproject/cortex/pkg/storage/backend/filesystem"
+	"github.com/cortexproject/cortex/pkg/storage/backend/gcs"
+	"github.com/cortexproject/cortex/pkg/storage/backend/s3"
 )
 
 // NewBucketClient creates a new bucket client based on the configured backend
